requirements.txt
Pillow>=10.3.0,<11.0.0

